Make sure the finalizer descriptor is allocated
Because of constructor ordering issues, the descriptor for finalizers
may not actually have been allocated before a finalizer is used. This
should only be a problem for GGGGC itself, since it's a constructor
within GGGGC; external constructors should not be able to create similar
dependency loops. Simply need to check that the finalizer descriptor is
allocated here.